+++ SBX/System/Systems/AttractorSystem.cpp
@@ -10,7 +10,7 @@ void AttractorSystem::Tick( float dt )
 {
 	for ( auto &accel : m_Components->Accelerations.GetContainer() )
 	{
-		constexpr double gravity = 3.0;
+		constexpr float gravity = 3.0;
 		accel.second.Acceleration += gravity * Utils::UnitVector( -m_Components->Positions.Find( accel.first ).Position );
 	}
 }
